Add 'edge-collapse' parameter to generate_clusters()
For now, I will not implement the 'directed' parameter, since
I need to run some tests on how to combine the edges first. This
'edge-collapse' parameter is a good way to test which kind of
undirected graph performs better.
Second I need to investigate whether there is an implementation
of Louvain for directed graphs.
Related to #15